test: factor out a test harness with parsing support
